Move i18n-data into common module.
