+++ common/i18n/proguard-rules.pro
