+++ i18n-domain/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="cz.kotox.i18n.domain" />
+<manifest package="cz.kotox.comon.i18n" />
