+++ i18n-domain/src/main/kotlin/cz/kotox/common/domain/usecase/CountryCodeLocalUseCase.kt
@@ -1,11 +1,11 @@
-package cz.kotox.domain.usecase
+package cz.kotox.common.domain.usecase
 
 import android.content.Context
 import android.telephony.TelephonyManager
-import cz.kotox.data.api.repository.CountryRepository
-import cz.kotox.domain.mapper.CountryModelMapper
-import cz.kotox.domain.model.CountryUiModel
-import cz.kotox.domain.model.CountryUiModelValueItem
+import cz.kotox.common.i18n.data.api.repository.CountryRepository
+import cz.kotox.common.domain.mapper.CountryModelMapper
+import cz.kotox.common.domain.model.CountryUiModel
+import cz.kotox.common.domain.model.CountryUiModelValueItem
 import dagger.hilt.android.qualifiers.ApplicationContext
 import java.util.Locale
 import javax.inject.Inject
